dev(utils): update to typescript
