+++ src/hooks/useCombobox/__tests__/getMenuProps.test.js
@@ -1,5 +1,5 @@
 import {act, renderHook} from '@testing-library/react'
-import {noop} from '../../../utils-ts'
+import {noop} from '../utils'
 import {getInput, renderCombobox, renderUseCombobox} from '../testUtils'
 import {
   defaultIds,
