+++ src/hooks/useCombobox/utils/getInitialState.js
@@ -1,0 +1,22 @@
+import {getInitialState as getInitialDropdownState} from '.'
+
+export function getInitialState(props) {
+  const initialState = getInitialDropdownState(props)
+  const {selectedItem} = initialState
+  let {inputValue} = initialState
+
+  if (
+    inputValue === '' &&
+    selectedItem &&
+    props.defaultInputValue === undefined &&
+    props.initialInputValue === undefined &&
+    props.inputValue === undefined
+  ) {
+    inputValue = props.itemToString(selectedItem)
+  }
+
+  return {
+    ...initialState,
+    inputValue,
+  }
+}
