+++ src/hooks/useSelect/utils/index.ts
@@ -1,3 +1,32 @@
+export {
+  useLatestRef,
+  validatePropTypes,
+  callAllEventHandlers,
+  handleRefs,
+  debounce,
+  normalizeArrowKey,
+  useGetterPropsCalledChecker,
+} from '../../../utils'
+export {
+  useControlPropsValidator,
+  isDropdownsStateEqual,
+  getItemAndIndex,
+  useControlledReducer,
+  useA11yMessageStatus,
+  isAcceptedCharacterKey,
+  useScrollIntoView,
+  getInitialValue,
+  getInitialState,
+} from '../../utils'
+export {
+  useElementIds,
+  defaultStateValues,
+  useMouseAndTouchTracker,
+  getHighlightedIndexOnOpen,
+  propTypes as dropdownPropTypes,
+  defaultProps as dropdownDefaultProps,
+} from '../../utils.dropdown'
+
 export {propTypes} from './propTypes'
 export {defaultProps} from './defaultProps'
 export {getItemIndexByCharacterKey} from './getItemIndexByCharacterKey'
