+++ src/hooks/utils.dropdown/index.ts
@@ -1,3 +1,6 @@
-export {propTypes as dropdownPropTypes} from './propTypes'
-export {defaultProps as dropdownDefaultProps} from './defaultProps'
-export {defaultStateValues as dropdownDefaultStateValues} from './defaultStateValues'
+export {propTypes} from './propTypes'
+export {defaultProps} from './defaultProps'
+export {defaultStateValues} from './defaultStateValues'
+export {useElementIds} from './useElementIds'
+export {getHighlightedIndexOnOpen} from './getHighlightedIndexOnOpen'
+export {useMouseAndTouchTracker} from './useMouseAndTouchTracker'
