+++ src/hooks/utils.dropdown/propTypes.ts
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 
-import {commonPropTypes} from '../utils-ts'
+import {commonPropTypes} from '../utils'
 
 // Shared between useSelect, useCombobox, Downshift.
 export const propTypes = {
