+++ src/hooks/utils/getInitialValue.ts
@@ -1,4 +1,4 @@
-import {State} from '../../utils-ts'
+import {State} from '../../utils'
 import {capitalizeString} from './capitalizeString'
 import {getDefaultValue} from './getDefaultValue'
