+++ src/hooks/utils/getItemAndIndex.ts
