+++ src/hooks/utils/isAcceptedCharacterKey.ts
@@ -1,0 +1,3 @@
+export function isAcceptedCharacterKey(key: string): boolean {
+  return /^\S{1}$/.test(key)
+}
