+++ src/index.types.ts
@@ -1,0 +1,6 @@
+export interface Environment {
+  addEventListener: typeof window.addEventListener
+  removeEventListener: typeof window.removeEventListener
+  document: Document
+  Node: typeof window.Node
+}
