+++ src/utils/index.ts
@@ -2,10 +2,13 @@ export {generateId, setIdCounter, resetIdCounter} from './generateId'
 export {useLatestRef} from './useLatestRef'
 export {handleRefs} from './handleRefs'
 export {callAllEventHandlers} from './callAllEventHandlers'
-export {debounce} from './debounce'
 export {setStatus, cleanupStatusDiv} from './setA11yStatus'
 export {noop} from './noop'
 export {validatePropTypes} from './validatePropTypes'
 export {getState} from './getState'
 export type {Action, Props, State} from './getState'
 export {scrollIntoView} from './scrollIntoView'
+export {useGetterPropsCalledChecker} from './useGetterPropsCalledChecker'
+export {debounce} from './debounce'
+
+export * from './index.legacy'
