+++ src/utils/noop.ts
